@@ -1,10 +1,10 @@
 module github.com/apache/pulsar-client-go/oauth2
 
-go 1.13
+go 1.15
 
 require (
-	github.com/99designs/keyring v1.1.6
-	github.com/form3tech-oss/jwt-go v3.2.3+incompatible
+	github.com/99designs/keyring v1.2.0
+	github.com/golang-jwt/jwt v3.2.2+incompatible
 	github.com/onsi/ginkgo v1.14.0
 	github.com/onsi/gomega v1.10.1
 	github.com/pkg/errors v0.9.1
