@@ -20,6 +20,7 @@ package pulsar
 import (
 	"context"
 	"fmt"
+	"io/ioutil"
 	"log"
 	"net/http"
 	"strconv"
@@ -27,8 +28,13 @@ import (
 	"testing"
 	"time"
 
+	"github.com/apache/pulsar-client-go/pulsar/crypto"
 	"github.com/apache/pulsar-client-go/pulsar/internal"
+	pb "github.com/apache/pulsar-client-go/pulsar/internal/pulsar_proto"
+	plog "github.com/apache/pulsar-client-go/pulsar/log"
+	"github.com/gogo/protobuf/proto"
 	"github.com/google/uuid"
+	"github.com/pierrec/lz4"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -90,7 +96,6 @@ func TestProducerConsumer(t *testing.T) {
 		assert.Equal(t, []byte(expectMsg), msg.Payload())
 		assert.Equal(t, "pulsar", msg.Key())
 		assert.Equal(t, expectProperties, msg.Properties())
-
 		// ack message
 		consumer.Ack(msg)
 	}
@@ -325,14 +330,77 @@ func TestPartitionTopicsConsumerPubSub(t *testing.T) {
 	require.NoError(t, err)
 	defer client.Close()
 
-	topic := "persistent://public/default/testGetPartitions"
-	testURL := adminURL + "/" + "admin/v2/persistent/public/default/testGetPartitions/partitions"
+	topic := "persistent://public/default/testGetPartitions5"
+	testURL := adminURL + "/" + "admin/v2/persistent/public/default/testGetPartitions5/partitions"
 
 	makeHTTPCall(t, http.MethodPut, testURL, "64")
 
 	// create producer
 	producer, err := client.CreateProducer(ProducerOptions{
 		Topic: topic,
+	})
+	assert.Nil(t, err)
+	defer producer.Close()
+
+	topics, err := client.TopicPartitions(topic)
+	assert.Nil(t, err)
+	assert.Equal(t, topic+"-partition-0", topics[0])
+	assert.Equal(t, topic+"-partition-1", topics[1])
+	assert.Equal(t, topic+"-partition-2", topics[2])
+
+	consumer, err := client.Subscribe(ConsumerOptions{
+		Topic:             topic,
+		SubscriptionName:  "my-sub",
+		Type:              Exclusive,
+		ReceiverQueueSize: 10,
+	})
+	assert.Nil(t, err)
+	defer consumer.Close()
+
+	ctx := context.Background()
+	for i := 0; i < 10; i++ {
+		_, err := producer.Send(ctx, &ProducerMessage{
+			Payload: []byte(fmt.Sprintf("hello-%d", i)),
+		})
+		assert.Nil(t, err)
+	}
+
+	msgs := make([]string, 0)
+
+	for i := 0; i < 10; i++ {
+		msg, err := consumer.Receive(ctx)
+		assert.Nil(t, err)
+		msgs = append(msgs, string(msg.Payload()))
+
+		fmt.Printf("Received message msgId: %#v -- content: '%s'\n",
+			msg.ID(), string(msg.Payload()))
+
+		consumer.Ack(msg)
+	}
+
+	assert.Equal(t, len(msgs), 10)
+}
+
+func TestPartitionTopicsConsumerPubSubEncryption(t *testing.T) {
+	client, err := NewClient(ClientOptions{
+		URL: lookupURL,
+	})
+	assert.Nil(t, err)
+	defer client.Close()
+
+	topic := "persistent://public/default/testGetPartitions"
+	testURL := adminURL + "/" + "admin/v2/persistent/public/default/testGetPartitions/partitions"
+
+	makeHTTPCall(t, http.MethodPut, testURL, "6")
+
+	// create producer
+	producer, err := client.CreateProducer(ProducerOptions{
+		Topic: topic,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"client-rsa.pem"},
+		},
 	})
 	require.NoError(t, err)
 	defer producer.Close()
@@ -348,6 +416,11 @@ func TestPartitionTopicsConsumerPubSub(t *testing.T) {
 		SubscriptionName:  "my-sub",
 		Type:              Exclusive,
 		ReceiverQueueSize: 10,
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			ConsumerCryptoFailureAction: crypto.ConsumerCryptoFailureActionFail,
+		},
 	})
 	require.NoError(t, err)
 	defer consumer.Close()
@@ -1414,6 +1487,107 @@ func TestRLQMultiTopics(t *testing.T) {
 	assert.Nil(t, checkMsg)
 }
 
+func TestRLQSpecifiedPartitionTopic(t *testing.T) {
+	topic := newTopicName()
+	testURL := adminURL + "/" + "admin/v2/persistent/public/default/" + topic + "/partitions"
+	makeHTTPCall(t, http.MethodPut, testURL, "1")
+
+	normalTopic := "persistent://public/default/" + topic
+	partitionTopic := normalTopic + "-partition-0"
+
+	subName := fmt.Sprintf("sub01-%d", time.Now().Unix())
+	maxRedeliveries := 2
+	N := 100
+	ctx := context.Background()
+
+	client, err := NewClient(ClientOptions{URL: lookupURL})
+	assert.Nil(t, err)
+	defer client.Close()
+
+	// subscribe topic with partition
+	rlqConsumer, err := client.Subscribe(ConsumerOptions{
+		Topic:                       partitionTopic,
+		SubscriptionName:            subName,
+		Type:                        Shared,
+		SubscriptionInitialPosition: SubscriptionPositionEarliest,
+		DLQ:                         &DLQPolicy{MaxDeliveries: uint32(maxRedeliveries)},
+		RetryEnable:                 true,
+		NackRedeliveryDelay:         1 * time.Second,
+	})
+	assert.Nil(t, err)
+	defer rlqConsumer.Close()
+
+	// subscribe DLQ Topic
+	dlqConsumer, err := client.Subscribe(ConsumerOptions{
+		Topic:                       subName + "-DLQ",
+		SubscriptionName:            subName,
+		SubscriptionInitialPosition: SubscriptionPositionEarliest,
+	})
+	assert.Nil(t, err)
+	defer dlqConsumer.Close()
+
+	// create producer
+	producer, err := client.CreateProducer(ProducerOptions{Topic: normalTopic})
+	assert.Nil(t, err)
+	defer producer.Close()
+
+	// 1. Pre-produce N messages
+	for i := 0; i < N; i++ {
+		_, err = producer.Send(ctx, &ProducerMessage{Payload: []byte(fmt.Sprintf("MSG_01_%d", i))})
+		assert.Nil(t, err)
+	}
+
+	// 2. Create consumer on the Retry Topics to reconsume N messages (maxRedeliveries+1) times
+	rlqReceived := 0
+	for rlqReceived < N*(maxRedeliveries+1) {
+		msg, err := rlqConsumer.Receive(ctx)
+		assert.Nil(t, err)
+		rlqConsumer.ReconsumeLater(msg, 1*time.Second)
+		rlqReceived++
+	}
+	fmt.Println("retry consumed:", rlqReceived) // 300
+
+	// No more messages on the Retry Topic
+	rlqCtx, rlqCancel := context.WithTimeout(context.Background(), 500*time.Millisecond)
+	defer rlqCancel()
+	msg, err := rlqConsumer.Receive(rlqCtx)
+	assert.Error(t, err)
+	assert.Nil(t, msg)
+
+	// 3. Create consumer on the DLQ topic to verify the routing
+	dlqReceived := 0
+	for dlqReceived < N {
+		msg, err := dlqConsumer.Receive(ctx)
+		assert.Nil(t, err)
+		dlqConsumer.Ack(msg)
+		dlqReceived++
+	}
+	fmt.Println("dlq received:", dlqReceived) // 100
+
+	// No more messages on the DLQ Topic
+	dlqCtx, dlqCancel := context.WithTimeout(context.Background(), 500*time.Millisecond)
+	defer dlqCancel()
+	msg, err = dlqConsumer.Receive(dlqCtx)
+	assert.Error(t, err)
+	assert.Nil(t, msg)
+
+	// 4. No more messages for same subscription
+	checkConsumer, err := client.Subscribe(ConsumerOptions{
+		Topic:                       normalTopic,
+		SubscriptionName:            subName,
+		Type:                        Shared,
+		SubscriptionInitialPosition: SubscriptionPositionEarliest,
+	})
+	assert.Nil(t, err)
+	defer checkConsumer.Close()
+
+	timeoutCtx, cancel := context.WithTimeout(context.Background(), 500*time.Millisecond)
+	defer cancel()
+	checkMsg, err := checkConsumer.Receive(timeoutCtx)
+	assert.Error(t, err)
+	assert.Nil(t, checkMsg)
+}
+
 func TestGetDeliveryCount(t *testing.T) {
 	client, err := NewClient(ClientOptions{
 		URL: lookupURL,
@@ -2063,4 +2237,844 @@ func TestConsumerKeySharedWithOrderingKey(t *testing.T) {
 		receivedConsumer1, receivedConsumer2,
 	)
 	assert.Equal(t, 100, receivedConsumer1+receivedConsumer2)
+}
+
+func TestProducerConsumerRSAEncryption(t *testing.T) {
+	client, err := NewClient(ClientOptions{
+		URL: lookupURL,
+	})
+	assert.Nil(t, err)
+	defer client.Close()
+
+	topic := fmt.Sprintf("my-topic-enc-%v", time.Now().Nanosecond())
+
+	cryptoConsumer, err := client.Subscribe(ConsumerOptions{
+		Topic: topic,
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			ConsumerCryptoFailureAction: crypto.ConsumerCryptoFailureActionFail,
+		},
+		SubscriptionName: "crypto-subscription",
+		Schema:           NewStringSchema(nil),
+	})
+
+	assert.Nil(t, err)
+
+	normalConsumer, err := client.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: "normal-subscription",
+		Schema:           NewStringSchema(nil),
+	})
+
+	assert.Nil(t, err)
+
+	cryptoProducer, err := client.CreateProducer(ProducerOptions{
+		Topic: topic,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"client-rsa.pem"},
+		},
+		Schema: NewStringSchema(nil),
+	})
+
+	assert.Nil(t, err)
+
+	msgFormat := "my-message-%v"
+
+	totalMessages := 10
+
+	ctx := context.Background()
+
+	for i := 0; i < totalMessages; i++ {
+		_, err := cryptoProducer.Send(ctx, &ProducerMessage{
+			Value: fmt.Sprintf(msgFormat, i),
+		})
+
+		assert.Nil(t, err)
+	}
+
+	// try to consume with normal consumer
+	normalConsumerCtx, cancel := context.WithTimeout(context.Background(), 500*time.Millisecond)
+	defer cancel()
+
+	msg, err := normalConsumer.Receive(normalConsumerCtx)
+	// msg should be null as the consumer will not be able to decrypt
+	assert.NotNil(t, err)
+	assert.Nil(t, msg)
+
+	// try to consume the message by crypto consumer
+	// consumer should be able to read all the messages
+	var actualMessage *string
+	for i := 0; i < totalMessages; i++ {
+		msg, err := cryptoConsumer.Receive(ctx)
+		fmt.Println(msg)
+		assert.Nil(t, err)
+		expectedMsg := fmt.Sprintf(msgFormat, i)
+		err = msg.GetSchemaValue(&actualMessage)
+		assert.Nil(t, err)
+		assert.Equal(t, expectedMsg, *actualMessage)
+		cryptoConsumer.Ack(msg)
+	}
+}
+
+func TestProducerConsumerRSAEncryptionWithCompression(t *testing.T) {
+	client, err := NewClient(ClientOptions{
+		URL: lookupURL,
+	})
+	assert.Nil(t, err)
+	defer client.Close()
+
+	topic := fmt.Sprintf("my-topic-enc-%v", time.Now().Nanosecond())
+
+	cryptoConsumer, err := client.Subscribe(ConsumerOptions{
+		Topic: topic,
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+		},
+		SubscriptionName: "crypto-subscription",
+		Schema:           NewStringSchema(nil),
+	})
+
+	assert.Nil(t, err)
+
+	normalConsumer, err := client.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: "normal-subscription",
+		Schema:           NewStringSchema(nil),
+	})
+
+	assert.Nil(t, err)
+
+	cryptoProducer, err := client.CreateProducer(ProducerOptions{
+		Topic: topic,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"client-rsa.pem"},
+		},
+		Schema:          NewStringSchema(nil),
+		CompressionType: LZ4,
+	})
+
+	assert.Nil(t, err)
+
+	msgFormat := "my-message-%v"
+
+	totalMessages := 10
+
+	ctx := context.Background()
+
+	for i := 0; i < totalMessages; i++ {
+		_, err := cryptoProducer.Send(ctx, &ProducerMessage{
+			Value: fmt.Sprintf(msgFormat, i),
+		})
+
+		assert.Nil(t, err)
+	}
+
+	// try to consume with normal consumer
+	normalConsumerCtx, cancel := context.WithTimeout(context.Background(), 500*time.Millisecond)
+	defer cancel()
+
+	msg, err := normalConsumer.Receive(normalConsumerCtx)
+	// msg should be null as the consumer will not be able to decrypt
+	assert.NotNil(t, err)
+	assert.Nil(t, msg)
+
+	// try to consume the message by crypto consumer
+	// consumer should be able to read all the messages
+	var actualMessage *string
+	for i := 0; i < totalMessages; i++ {
+		msg, err := cryptoConsumer.Receive(ctx)
+		assert.Nil(t, err)
+		expectedMsg := fmt.Sprintf(msgFormat, i)
+		err = msg.GetSchemaValue(&actualMessage)
+		assert.Nil(t, err)
+		assert.Equal(t, expectedMsg, *actualMessage)
+		cryptoConsumer.Ack(msg)
+	}
+}
+
+func TestBatchProducerConsumerRSAEncryptionWithCompression(t *testing.T) {
+	client, err := NewClient(ClientOptions{
+		URL: lookupURL,
+	})
+	assert.Nil(t, err)
+	defer client.Close()
+
+	topic := fmt.Sprintf("my-topic-enc-%v", time.Now().Nanosecond())
+
+	cryptoConsumer, err := client.Subscribe(ConsumerOptions{
+		Topic: topic,
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+		},
+		SubscriptionName: "crypto-subscription",
+		Schema:           NewStringSchema(nil),
+	})
+
+	assert.Nil(t, err)
+
+	normalConsumer, err := client.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: "normal-subscription",
+		Schema:           NewStringSchema(nil),
+	})
+
+	assert.Nil(t, err)
+	batchSize := 2
+	cryptoProducer, err := client.CreateProducer(ProducerOptions{
+		Topic: topic,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"client-rsa.pem"},
+		},
+		Schema:              NewStringSchema(nil),
+		CompressionType:     LZ4,
+		DisableBatching:     false,
+		BatchingMaxMessages: uint(batchSize),
+	})
+
+	assert.Nil(t, err)
+
+	msgFormat := "my-message-%v"
+
+	totalMessages := 10
+
+	ctx := context.Background()
+
+	for i := 0; i < totalMessages; i++ {
+		_, err := cryptoProducer.Send(ctx, &ProducerMessage{
+			Value: fmt.Sprintf(msgFormat, i),
+		})
+
+		assert.Nil(t, err)
+	}
+
+	// try to consume with normal consumer
+	normalConsumerCtx, cancel := context.WithTimeout(context.Background(), 500*time.Millisecond)
+	defer cancel()
+
+	msg, err := normalConsumer.Receive(normalConsumerCtx)
+	// msg should be null as the consumer will not be able to decrypt
+	assert.NotNil(t, err)
+	assert.Nil(t, msg)
+
+	// try to consume the message by crypto consumer
+	// consumer should be able to read all the messages
+	var actualMessage *string
+	for i := 0; i < totalMessages; i++ {
+		msg, err := cryptoConsumer.Receive(ctx)
+		assert.Nil(t, err)
+		expectedMsg := fmt.Sprintf(msgFormat, i)
+		err = msg.GetSchemaValue(&actualMessage)
+		assert.Nil(t, err)
+		assert.Equal(t, expectedMsg, *actualMessage)
+		cryptoConsumer.Ack(msg)
+	}
+}
+
+func TestProducerConsumerRedeliveryOfFailedEncryptedMessages(t *testing.T) {
+	// create new client instance for each producer and consumer
+	client, err := NewClient(ClientOptions{
+		URL: serviceURL,
+	})
+	assert.Nil(t, err)
+	defer client.Close()
+
+	clientCryptoConsumer, err := NewClient(ClientOptions{
+		URL: serviceURL,
+	})
+	assert.Nil(t, err)
+	defer clientCryptoConsumer.Close()
+
+	clientCryptoConsumerInvalidKeyReader, err := NewClient(ClientOptions{
+		URL: serviceURL,
+	})
+	assert.Nil(t, err)
+	defer clientCryptoConsumerInvalidKeyReader.Close()
+
+	clientcryptoConsumerNoKeyReader, err := NewClient(ClientOptions{
+		URL: serviceURL,
+	})
+	assert.Nil(t, err)
+	defer clientcryptoConsumerNoKeyReader.Close()
+
+	topic := fmt.Sprintf("my-topic-enc-%v", time.Now().Nanosecond())
+
+	cryptoProducer, err := client.CreateProducer(ProducerOptions{
+		Topic: topic,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"client-rsa.pem"},
+		},
+		CompressionType: LZ4,
+		Schema:          NewStringSchema(nil),
+	})
+	assert.Nil(t, err)
+
+	sharedSubscription := "crypto-shared-subscription"
+
+	cryptoConsumer, err := clientCryptoConsumer.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: sharedSubscription,
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+		},
+		Schema:              NewStringSchema(nil),
+		Type:                Shared,
+		NackRedeliveryDelay: 1 * time.Second,
+	})
+	assert.Nil(t, err)
+
+	cryptoConsumerInvalidKeyReader, err := clientCryptoConsumerInvalidKeyReader.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: sharedSubscription,
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa_invalid.pem"),
+		},
+		Schema:              NewStringSchema(nil),
+		Type:                Shared,
+		NackRedeliveryDelay: 1 * time.Second,
+	})
+	assert.Nil(t, err)
+
+	cryptoConsumerNoKeyReader, err := clientcryptoConsumerNoKeyReader.Subscribe(ConsumerOptions{
+		Topic:               topic,
+		SubscriptionName:    sharedSubscription,
+		Schema:              NewStringSchema(nil),
+		Type:                Shared,
+		NackRedeliveryDelay: 1 * time.Second,
+	})
+	assert.Nil(t, err)
+
+	totalMessages := 5
+	message := "my-message-%v"
+	// since messages can be in random order
+	// map can be used to check if all the messages are received
+	messageMap := map[string]struct{}{}
+
+	// producer messages
+	for i := 0; i < totalMessages; i++ {
+		mid, err := cryptoProducer.Send(context.Background(), &ProducerMessage{
+			Value: fmt.Sprintf(message, i),
+		})
+		assert.Nil(t, err)
+		fmt.Printf("Sent : %v\n", mid)
+	}
+
+	// Consuming from consumer 2 and 3
+	// no message should be returned since they can't decrypt the message
+	ctxWithTimeOut1, c1 := context.WithTimeout(context.Background(), 2*time.Second)
+	defer c1()
+
+	ctxWithTimeOut2, c2 := context.WithTimeout(context.Background(), 2*time.Second)
+	defer c2()
+
+	// try to consume messages
+	msg, err := cryptoConsumerInvalidKeyReader.Receive(ctxWithTimeOut1)
+	assert.NotNil(t, err)
+	assert.Nil(t, msg)
+
+	msg, err = cryptoConsumerNoKeyReader.Receive(ctxWithTimeOut2)
+	assert.NotNil(t, err)
+	assert.Nil(t, msg)
+
+	cryptoConsumerInvalidKeyReader.Close()
+	cryptoConsumerNoKeyReader.Close()
+
+	// try to consume by consumer1
+	// all the messages would by received by it
+	var receivedMsg *string
+	for i := 0; i < totalMessages; i++ {
+		m, err := cryptoConsumer.Receive(context.Background())
+		assert.Nil(t, err)
+		err = m.GetSchemaValue(&receivedMsg)
+		assert.Nil(t, err)
+		messageMap[*receivedMsg] = struct{}{}
+		cryptoConsumer.Ack(m)
+		fmt.Printf("Received : %v\n", m.ID())
+	}
+
+	// check if all messages were received
+	for i := 0; i < totalMessages; i++ {
+		key := fmt.Sprintf(message, i)
+		_, ok := messageMap[key]
+		assert.True(t, ok)
+	}
+}
+
+func TestRSAEncryptionFailure(t *testing.T) {
+	client, err := NewClient(ClientOptions{
+		URL: serviceURL,
+	})
+	assert.Nil(t, err)
+	client.Close()
+
+	topic := fmt.Sprintf("my-topic-enc-%v", time.Now().Nanosecond())
+
+	// 1. invalid key name
+	// create producer with invalid key
+	// producer creation succeeds but message sending should fail with an error
+	producer, err := client.CreateProducer(ProducerOptions{
+		Topic: topic,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa_invalid.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"client-rsa.pem"},
+		},
+		Schema:          NewStringSchema(nil),
+		DisableBatching: true,
+	})
+	assert.Nil(t, err)
+	assert.NotNil(t, producer)
+
+	// sending of message should fail with an error, since invalid rsa keys are configured
+	mid, err := producer.Send(context.Background(), &ProducerMessage{
+		Value: "some-message",
+	})
+
+	assert.Nil(t, mid)
+	assert.NotNil(t, err)
+	producer.Close()
+
+	// 2. Producer with valid key name
+	producer, err = client.CreateProducer(ProducerOptions{
+		Topic: topic,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"client-rsa.pem"},
+		},
+		Schema:          NewStringSchema(nil),
+		DisableBatching: true,
+	})
+	assert.Nil(t, err)
+	assert.NotNil(t, producer)
+
+	subscriptionName := "enc-failure-subcription"
+	totalMessages := 10
+
+	consumer, err := client.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: subscriptionName,
+	})
+	assert.Nil(t, err)
+
+	messageFormat := "my-message-%v"
+	for i := 0; i < totalMessages; i++ {
+		_, err := producer.Send(context.Background(), &ProducerMessage{
+			Value: fmt.Sprintf(messageFormat, i),
+		})
+		assert.Nil(t, err)
+	}
+
+	// 3. KeyReader is not set by the consumer
+	// Receive should fail since KeyReader is not setup
+	// because default behaviour of consumer is fail receiving message if error in decryption
+	ctx, cancel := context.WithTimeout(context.Background(), 2*time.Second)
+	defer cancel()
+
+	msg, err := consumer.Receive(ctx)
+	assert.NotNil(t, err)
+	assert.Nil(t, msg, "Receive should have failed with no keyreader")
+
+	// 4. Set consumer config to consume even if decryption fails
+	consumer.Close()
+	consumer, err = client.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: subscriptionName,
+		Decryption: &MessageDecryptionInfo{
+			ConsumerCryptoFailureAction: crypto.ConsumerCryptoFailureActionConsume,
+		},
+		Schema: NewStringSchema(nil),
+	})
+	assert.Nil(t, err)
+	assert.NotNil(t, consumer)
+
+	ctx2, cancel2 := context.WithTimeout(context.Background(), 2*time.Second)
+	defer cancel2()
+
+	for i := 0; i < totalMessages-1; i++ {
+		expectedMessage := fmt.Sprintf(messageFormat, i)
+		msg, err = consumer.Receive(ctx2)
+		assert.Nil(t, err)
+		assert.NotNil(t, msg)
+
+		receivedMsg := string(msg.Payload())
+		assert.NotEqual(t, expectedMessage, receivedMsg, fmt.Sprintf(`Received encrypted message [%v]
+	should not match the expected message  [%v]`, expectedMessage, receivedMsg))
+		// verify the message contains Encryption context
+		assert.NotEmpty(t, msg.GetEncryptionContext(),
+			"Encrypted message which is failed to decrypt must contain EncryptionContext")
+		consumer.Ack(msg)
+	}
+
+	// 5. discard action on decryption failure
+	consumer.Close()
+	consumer, err = client.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: subscriptionName,
+		Decryption: &MessageDecryptionInfo{
+			ConsumerCryptoFailureAction: crypto.ConsumerCryptoFailureActionDiscard,
+		},
+		Schema: NewStringSchema(nil),
+	})
+	assert.Nil(t, err)
+	assert.NotNil(t, consumer)
+
+	ctx3, cancel3 := context.WithTimeout(context.Background(), 3*time.Second)
+	defer cancel3()
+
+	msg, err = consumer.Receive(ctx3)
+	assert.NotNil(t, err)
+	assert.Nil(t, msg, "Message received even aftet ConsumerCryptoFailureAction.Discard is set.")
+}
+
+func TestConsumerCompressionWithRSAEncryption(t *testing.T) {
+	client, err := NewClient(ClientOptions{
+		URL: lookupURL,
+	})
+
+	assert.Nil(t, err)
+	defer client.Close()
+
+	topicName := newTopicName()
+	ctx := context.Background()
+
+	producer, err := client.CreateProducer(ProducerOptions{
+		Topic:           topicName,
+		CompressionType: LZ4,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"enc-compress-app.key"},
+		},
+	})
+
+	assert.Nil(t, err)
+	defer producer.Close()
+
+	consumer, err := client.Subscribe(ConsumerOptions{
+		Topic:            topicName,
+		SubscriptionName: "sub-1",
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+		},
+	})
+
+	assert.Nil(t, err)
+	defer consumer.Close()
+
+	const N = 100
+
+	for i := 0; i < N; i++ {
+		if _, err := producer.Send(ctx, &ProducerMessage{
+			Payload: []byte(fmt.Sprintf("msg-content-%d", i)),
+		}); err != nil {
+			t.Fatal(err)
+		}
+	}
+
+	for i := 0; i < N; i++ {
+		msg, err := consumer.Receive(ctx)
+		assert.Nil(t, err)
+		assert.Equal(t, fmt.Sprintf("msg-content-%d", i), string(msg.Payload()))
+		consumer.Ack(msg)
+	}
+}
+
+func TestBatchMessageReceiveWithCompressionAndRSAEcnryption(t *testing.T) {
+	client, err := NewClient(ClientOptions{
+		URL: lookupURL,
+	})
+
+	assert.Nil(t, err)
+	defer client.Close()
+
+	topicName := "persistent://public/default/receive-batch-comp-enc"
+	subName := "subscription-name"
+	prefix := "msg-batch-"
+	ctx := context.Background()
+
+	// Enable batching on producer side
+	batchSize, numOfMessages := 2, 100
+
+	// create producer
+	producer, err := client.CreateProducer(ProducerOptions{
+		Topic:               topicName,
+		BatchingMaxMessages: uint(batchSize),
+		DisableBatching:     false,
+		CompressionType:     LZ4,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"batch-encryption-app.key"},
+		},
+	})
+	assert.Nil(t, err)
+	assert.Equal(t, topicName, producer.Topic())
+	defer producer.Close()
+
+	consumer, err := client.Subscribe(ConsumerOptions{
+		Topic:            topicName,
+		SubscriptionName: subName,
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: crypto.NewFileKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+		},
+	})
+
+	assert.Nil(t, err)
+	defer consumer.Close()
+
+	count := 0
+	for i := 0; i < numOfMessages; i++ {
+		messageContent := prefix + fmt.Sprintf("%d", i)
+		msg := &ProducerMessage{
+			Payload: []byte(messageContent),
+		}
+		_, err := producer.Send(ctx, msg)
+		assert.Nil(t, err)
+	}
+
+	for i := 0; i < numOfMessages; i++ {
+		msg, err := consumer.Receive(ctx)
+		fmt.Printf("received : %v\n", string(msg.Payload()))
+		assert.Nil(t, err)
+		consumer.Ack(msg)
+		count++
+	}
+
+	assert.Equal(t, count, numOfMessages)
+}
+
+type EncKeyReader struct {
+	publicKeyPath  string
+	privateKeyPath string
+	metaMap        map[string]string
+}
+
+func NewEncKeyReader(publicKeyPath, privateKeyPath string) *EncKeyReader {
+	metaMap := map[string]string{
+		"version": "1.0",
+	}
+
+	return &EncKeyReader{
+		publicKeyPath:  publicKeyPath,
+		privateKeyPath: privateKeyPath,
+		metaMap:        metaMap,
+	}
+}
+
+// GetPublicKey read public key from the given path
+func (d *EncKeyReader) PublicKey(keyName string, keyMeta map[string]string) (*crypto.EncryptionKeyInfo, error) {
+	return readKey(keyName, d.publicKeyPath, d.metaMap)
+}
+
+// GetPrivateKey read private key from the given path
+func (d *EncKeyReader) PrivateKey(keyName string, keyMeta map[string]string) (*crypto.EncryptionKeyInfo, error) {
+	return readKey(keyName, d.privateKeyPath, d.metaMap)
+}
+
+func readKey(keyName, path string, keyMeta map[string]string) (*crypto.EncryptionKeyInfo, error) {
+	key, err := ioutil.ReadFile(path)
+	if err != nil {
+		return nil, err
+	}
+	return crypto.NewEncryptionKeyInfo(keyName, key, keyMeta), nil
+}
+
+func TestConsumerEncryptionWithoutKeyReader(t *testing.T) {
+	client, err := NewClient(ClientOptions{
+		URL: serviceURL,
+	})
+	assert.Nil(t, err)
+	defer client.Close()
+
+	topic := newTopicName()
+	encryptionKeyName := "client-rsa.pem"
+
+	producer, err := client.CreateProducer(ProducerOptions{
+		Topic: topic,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: NewEncKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{encryptionKeyName},
+		},
+		CompressionType: LZ4,
+		Schema:          NewStringSchema(nil),
+	})
+
+	assert.Nil(t, err)
+	assert.NotNil(t, producer)
+
+	consumer, err := client.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: "my-subscription-name",
+		Decryption: &MessageDecryptionInfo{
+			ConsumerCryptoFailureAction: crypto.ConsumerCryptoFailureActionConsume,
+		},
+		Schema: NewStringSchema(nil),
+	})
+	assert.Nil(t, err)
+
+	message := "my-message"
+
+	_, err = producer.Send(context.Background(), &ProducerMessage{
+		Value: message,
+	})
+	assert.Nil(t, err)
+
+	// consume encrypted message
+	msg, err := consumer.Receive(context.Background())
+	assert.Nil(t, err)
+	assert.NotNil(t, msg)
+
+	// try to decrypt message
+	encCtx := msg.GetEncryptionContext()
+	assert.NotEmpty(t, encCtx)
+
+	keys := encCtx.Keys
+	assert.Equal(t, 1, len(keys))
+
+	encryptionKey, ok := keys[encryptionKeyName]
+	assert.True(t, ok)
+
+	encDataKey := encryptionKey.KeyValue
+	assert.NotNil(t, encDataKey)
+
+	metadata := encryptionKey.Metadata
+	assert.NotNil(t, metadata)
+
+	version := metadata["version"]
+	assert.Equal(t, "1.0", version)
+
+	compressionType := encCtx.CompressionType
+	uncompressedSize := uint32(encCtx.UncompressedSize)
+	encParam := encCtx.Param
+	encAlgo := encCtx.Algorithm
+	batchSize := encCtx.BatchSize
+
+	// try to decrypt using default MessageCrypto
+	msgCrypto, err := crypto.NewDefaultMessageCrypto("testing", false, plog.DefaultNopLogger())
+	assert.Nil(t, err)
+
+	producerName := "test"
+	sequenceID := uint64(123)
+	publishTime := uint64(12333453454)
+
+	messageMetaData := pb.MessageMetadata{
+		EncryptionParam:  encParam,
+		ProducerName:     &producerName,
+		SequenceId:       &sequenceID,
+		PublishTime:      &publishTime,
+		UncompressedSize: &uncompressedSize,
+		EncryptionAlgo:   &encAlgo,
+	}
+
+	if compressionType == LZ4 {
+		messageMetaData.Compression = pb.CompressionType_LZ4.Enum()
+	}
+
+	messageMetaData.EncryptionKeys = []*pb.EncryptionKeys{{
+		Key:   &encryptionKeyName,
+		Value: encDataKey,
+	}}
+
+	decryptedPayload, err := msgCrypto.Decrypt(crypto.NewMessageMetadataSupplier(&messageMetaData),
+		msg.Payload(),
+		NewEncKeyReader("crypto/testdata/pub_key_rsa.pem",
+			"crypto/testdata/pri_key_rsa.pem"))
+	assert.Nil(t, err)
+	assert.NotNil(t, decryptedPayload)
+
+	// try to uncompress payload
+	uncompressedPayload := make([]byte, uncompressedSize)
+	s, err := lz4.UncompressBlock(decryptedPayload, uncompressedPayload)
+	assert.Nil(t, err)
+	assert.Equal(t, uncompressedSize, uint32(s))
+
+	buffer := internal.NewBufferWrapper(uncompressedPayload)
+
+	if batchSize > 0 {
+		size := buffer.ReadUint32()
+		var meta pb.SingleMessageMetadata
+		if err := proto.Unmarshal(buffer.Read(size), &meta); err != nil {
+			fmt.Println(err)
+		}
+		d := buffer.Read(uint32(meta.GetPayloadSize()))
+		assert.Equal(t, message, string(d))
+	}
+}
+
+// TestEncryptDecryptRedeliveryOnFailure test redelivery failed messages
+func TestEncryptDecryptRedeliveryOnFailure(t *testing.T) {
+	client, err := NewClient(ClientOptions{
+		URL: serviceURL,
+	})
+	assert.Nil(t, err)
+
+	topic := newTopicName()
+	subcription := "test-subscription-redelivery"
+
+	consumer, err := client.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: subcription,
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: NewEncKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_invalid_rsa.pem"),
+		},
+	})
+	assert.Nil(t, err)
+
+	producer, err := client.CreateProducer(ProducerOptions{
+		Topic: topic,
+		Encryption: &ProducerEncryptionInfo{
+			KeyReader: NewEncKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+			Keys: []string{"new-enc-key"},
+		},
+	})
+	assert.Nil(t, err)
+
+	producer.Send(context.Background(), &ProducerMessage{
+		Payload: []byte("new-test-message"),
+	})
+
+	ctx, cancel := context.WithTimeout(context.Background(), 1000*time.Millisecond)
+	defer cancel()
+
+	// message receive should fail due to decryption error
+	msg, err := consumer.Receive(ctx)
+	assert.Nil(t, msg)
+	assert.NotNil(t, err)
+
+	consumer.Close()
+
+	// create consumer with same subscription and proper rsa key pairs
+	consumer, err = client.Subscribe(ConsumerOptions{
+		Topic:            topic,
+		SubscriptionName: subcription,
+		Decryption: &MessageDecryptionInfo{
+			KeyReader: NewEncKeyReader("crypto/testdata/pub_key_rsa.pem",
+				"crypto/testdata/pri_key_rsa.pem"),
+		},
+	})
+	assert.Nil(t, err)
+
+	// previous message should be redelivered
+	msg, err = consumer.Receive(context.Background())
+	assert.Nil(t, err)
+	assert.NotNil(t, msg)
+	consumer.Ack(msg)
 }
