@@ -89,8 +89,7 @@ func (r *dlqRouter) run() {
 		select {
 		case cm := <-r.messageCh:
 			r.log.WithField("msgID", cm.ID()).Debug("Got message for DLQ")
-			producer := r.getProducer()
-
+			producer := r.getProducer(cm.Consumer.(*consumer).options.Schema)
 			msg := cm.Message.(*message)
 			msgID := msg.ID()
 			producer.SendAsync(context.Background(), &ProducerMessage{
@@ -127,7 +126,7 @@ func (r *dlqRouter) close() {
 	}
 }
 
-func (r *dlqRouter) getProducer() Producer {
+func (r *dlqRouter) getProducer(schema Schema) Producer {
 	if r.producer != nil {
 		// Producer was already initialized
 		return r.producer
@@ -140,6 +139,7 @@ func (r *dlqRouter) getProducer() Producer {
 			Topic:                   r.policy.DeadLetterTopic,
 			CompressionType:         LZ4,
 			BatchingMaxPublishDelay: 100 * time.Millisecond,
+			Schema:                  schema,
 		})
 
 		if err != nil {
