@@ -80,9 +80,10 @@ func (p *connectionPool) GetConnection(logicalAddr *url.URL, physicalAddr *url.U
 
 		// remove stale/failed connection
 		if conn.closed() {
-			delete(p.connections, key)
-			p.log.Debugf("Removed connection from pool key=%s logical_addr=%+v physical_addr=%+v",
+			p.log.Infof("Removed connection from pool key=%s logical_addr=%+v physical_addr=%+v",
 				key, conn.logicalAddr, conn.physicalAddr)
+			delete(p.connections, key)
+			conn.Close()
 			conn = nil // set to nil so we create a new one
 		}
 	}
