@@ -26,7 +26,7 @@ type ProducerMessage struct {
 	// Payload for the message
 	Payload []byte
 
-	//Value and payload is mutually exclusive, `Value interface{}` for schema message.
+	// Value and payload is mutually exclusive, `Value interface{}` for schema message.
 	Value interface{}
 
 	// Key sets the key of the message for routing policy
@@ -48,19 +48,19 @@ type ProducerMessage struct {
 	// ReplicationClusters override the replication clusters for this message.
 	ReplicationClusters []string
 
-	// Disable the replication for this message
+	// DisableReplication disables the replication for this message
 	DisableReplication bool
 
-	// SequenceID set the sequence id to assign to the current message
+	// SequenceID sets the sequence id to assign to the current message
 	SequenceID *int64
 
-	// Request to deliver the message only after the specified relative delay.
+	// DeliverAfter requests to deliver the message only after the specified relative delay.
 	// Note: messages are only delivered with delay when a consumer is consuming
 	//     through a `SubscriptionType=Shared` subscription. With other subscription
 	//     types, the messages will still be delivered immediately.
 	DeliverAfter time.Duration
 
-	// Deliver the message only at or after the specified absolute timestamp.
+	// DeliverAt delivers the message only at or after the specified absolute timestamp.
 	// Note: messages are only delivered with delay when a consumer is consuming
 	//     through a `SubscriptionType=Shared` subscription. With other subscription
 	//     types, the messages will still be delivered immediately.
@@ -69,53 +69,66 @@ type ProducerMessage struct {
 
 // Message abstraction used in Pulsar
 type Message interface {
-	// Topic get the topic from which this message originated from
+	// Topic returns the topic from which this message originated from.
 	Topic() string
 
 	// ProducerName returns the name of the producer that has published the message.
 	ProducerName() string
 
 	// Properties are application defined key/value pairs that will be attached to the message.
-	// Return the properties attached to the message.
+	// Returns the properties attached to the message.
 	Properties() map[string]string
 
-	// Payload get the payload of the message
+	// Payload returns the payload of the message
 	Payload() []byte
 
-	// ID get the unique message ID associated with this message.
+	// ID returns the unique message ID associated with this message.
 	// The message id can be used to univocally refer to a message without having the keep the entire payload in memory.
 	ID() MessageID
 
-	// PublishTime get the publish time of this message. The publish time is the timestamp that a client
+	// PublishTime returns the publish time of this message. The publish time is the timestamp that a client
 	// publish the message.
 	PublishTime() time.Time
 
-	// EventTime get the event time associated with this message. It is typically set by the applications via
+	// EventTime returns the event time associated with this message. It is typically set by the applications via
 	// `ProducerMessage.EventTime`.
 	// If EventTime is 0, it means there isn't any event time associated with this message.
 	EventTime() time.Time
 
-	// Key get the key of the message, if any
+	// Key returns the key of the message, if any
 	Key() string
 
-	// OrderingKey get the ordering key of the message, if any
+	// OrderingKey returns the ordering key of the message, if any
 	OrderingKey() string
 
-	// Get message redelivery count, redelivery count maintain in pulsar broker. When client nack acknowledge messages,
+	// RedeliveryCount returns message redelivery count, redelivery count maintain in pulsar broker.
+	// When client nack acknowledge messages,
 	// broker will dispatch message again with message redelivery count in CommandMessage defined.
 	//
 	// Message redelivery increases monotonically in a broker, when topic switch ownership to a another broker
 	// redelivery count will be recalculated.
 	RedeliveryCount() uint32
 
-	// Check whether the message is replicated from other cluster.
+	// IsReplicated determines whether the message is replicated from another cluster.
 	IsReplicated() bool
 
-	// Get name of cluster, from which the message is replicated.
+	// GetReplicatedFrom returns the name of the cluster, from which the message is replicated.
 	GetReplicatedFrom() string
 
-	//Get the de-serialized value of the message, according the configured
+	// GetSchemaValue returns the de-serialized value of the message, according to the configuration.
 	GetSchemaValue(v interface{}) error
+
+	// GetEncryptionContext returns the ecryption context of the message.
+	// It will be used by the application to parse the undecrypted message.
+	GetEncryptionContext() *EncryptionContext
+
+	// Index returns index from broker entry metadata,
+	// or empty if the feature is not enabled in the broker.
+	Index() *uint64
+
+	// BrokerPublishTime returns broker publish time from broker entry metadata,
+	// or empty if the feature is not enabled in the broker.
+	BrokerPublishTime() *time.Time
 }
 
 // MessageID identifier for a particular message
@@ -123,16 +136,16 @@ type MessageID interface {
 	// Serialize the message id into a sequence of bytes that can be stored somewhere else
 	Serialize() []byte
 
-	// Get the message ledgerID
+	// LedgerID returns the message ledgerID
 	LedgerID() int64
 
-	// Get the message entryID
+	// EntryID returns the message entryID
 	EntryID() int64
 
-	// Get the message batchIdx
+	// BatchIdx returns the message batchIdx
 	BatchIdx() int32
 
-	// Get the message partitionIdx
+	// PartitionIdx returns the message partitionIdx
 	PartitionIdx() int32
 }
 
@@ -146,7 +159,7 @@ func EarliestMessageID() MessageID {
 	return earliestMessageID
 }
 
-// LatestMessage returns a messageID that points to the latest message
+// LatestMessageID returns a messageID that points to the latest message
 func LatestMessageID() MessageID {
 	return latestMessageID
 }
