@@ -22,23 +22,23 @@ import (
 	"time"
 )
 
-// ReaderMessage package Reader and Message as a struct to use
+// ReaderMessage packages Reader and Message as a struct to use.
 type ReaderMessage struct {
 	Reader
 	Message
 }
 
-// ReaderOptions abstraction Reader options to use.
+// ReaderOptions represents Reader options to use.
 type ReaderOptions struct {
-	// Topic specify the topic this consumer will subscribe on.
+	// Topic specifies the topic this consumer will subscribe on.
 	// This argument is required when constructing the reader.
 	Topic string
 
 	// Name set the reader name.
 	Name string
 
-	// Attach a set of application defined properties to the reader
-	// This properties will be visible in the topic stats
+	// Properties represents a set of application defined properties for the reader.
+	// Those properties will be visible in the topic stats.
 	Properties map[string]string
 
 	// StartMessageID initial reader positioning is done by specifying a message id. The options are:
@@ -50,7 +50,7 @@ type ReaderOptions struct {
 	//                  messageID
 	StartMessageID MessageID
 
-	// If true, the reader will start at the `StartMessageID`, included.
+	// StartMessageIDInclusive, if true, the reader will start at the `StartMessageID`, included.
 	// Default is `false` and the reader will start from the "next" message
 	StartMessageIDInclusive bool
 
@@ -65,17 +65,27 @@ type ReaderOptions struct {
 	// Default value is {@code 1000} messages and should be good for most use cases.
 	ReceiverQueueSize int
 
-	// SubscriptionRolePrefix set the subscription role prefix. The default prefix is "reader".
+	// SubscriptionRolePrefix sets the subscription role prefix. The default prefix is "reader".
 	SubscriptionRolePrefix string
 
-	// If enabled, the reader will read messages from the compacted topic rather than reading the full message backlog
-	// of the topic. This means that, if the topic has been compacted, the reader will only see the latest value for
-	// each key in the topic, up until the point in the topic message backlog that has been compacted. Beyond that
-	// point, the messages will be sent as normal.
+	// SubscriptionName sets the subscription name.
+	// If subscriptionRolePrefix is set at the same time, this configuration will prevail
+	SubscriptionName string
+
+	// ReadCompacted, if enabled, the reader will read messages from the compacted topic rather than reading the
+	// full message backlog of the topic. This means that, if the topic has been compacted, the reader will only
+	// see the latest value for each key in the topic, up until the point in the topic message backlog that has
+	// been compacted. Beyond that point, the messages will be sent as normal.
 	//
 	// ReadCompacted can only be enabled when reading from a persistent topic. Attempting to enable it on non-persistent
 	// topics will lead to the reader create call throwing a PulsarClientException.
 	ReadCompacted bool
+
+	// Decryption represents the encryption related fields required by the reader to decrypt a message.
+	Decryption *MessageDecryptionInfo
+
+	// Schema represents the schema implementation.
+	Schema Schema
 }
 
 // Reader can be used to scan through all the messages currently available in a topic.
@@ -83,23 +93,23 @@ type Reader interface {
 	// Topic from which this reader is reading from
 	Topic() string
 
-	// Next read the next message in the topic, blocking until a message is available
+	// Next reads the next message in the topic, blocking until a message is available
 	Next(context.Context) (Message, error)
 
-	// HasNext check if there is any message available to read from the current position
+	// HasNext checks if there is any message available to read from the current position
 	HasNext() bool
 
 	// Close the reader and stop the broker to push more messages
 	Close()
 
-	// Reset the subscription associated with this reader to a specific message id.
+	// Seek resets the subscription associated with this reader to a specific message id.
 	// The message id can either be a specific message or represent the first or last messages in the topic.
 	//
 	// Note: this operation can only be done on non-partitioned topics. For these, one can rather perform the
 	//       seek() on the individual partitions.
 	Seek(MessageID) error
 
-	// Reset the subscription associated with this reader to a specific message publish time.
+	// SeekByTime resets the subscription associated with this reader to a specific message publish time.
 	//
 	// Note: this operation can only be done on non-partitioned topics. For these, one can rather perform the seek() on
 	// the individual partitions.
